refactor: tighten the public API and drop dead surface
Take the compression level as `u8` instead of `i32` in `pack_level` and
`workmem_size_level`. The valid range is 1 to 10, so a signed 32-bit
integer let most values stand for errors. `u8` rules out negatives and
large values while the range check still rejects 0 and 11 and up.

Drop the four `VER_*` constants. They stated a wire-format version that
the codec bytes do not carry. The format has no header or version field,
so a public version number on the crate had no meaning and disagreed with
the Cargo version anyway.

Remove `Error::as_sentinel`. Both variants returned the same all-ones
value, so the method carried no information. The typed `Result` already
models failure. A C wrapper, if ever built, can map errors at its own
boundary.

Drop the unused `src_size` parameter from `workmem_size`. The level-1
scratch size is fixed, so the parameter only invited callers to compute a
value with no effect.

Rewrite the doc comments that justified the crate by pointing at an
outside C codebase. Each one now states the Rust behavior directly.

+++ tests/common/mod.rs
@@ -108,10 +108,10 @@ pub const ERRORS: &[ErrorCase] = &[
     },
 ];
 
-/// Deterministic byte source matching the golden generator.
+/// Deterministic byte source for the round-trip and fuzz tests.
 ///
-/// A small linear congruential generator. The Rust tests and the captured
-/// golden data agree on these bytes, so the encoder check stays portable.
+/// A small linear congruential generator. A fixed seed yields the same bytes
+/// every run, so the random-shaped tests stay reproducible.
 pub struct Lcg(u32);
 
 impl Lcg {
@@ -135,7 +135,7 @@ impl Lcg {
 }
 
 /// Compress `src` at `level` into a fresh buffer and return the packed bytes.
-pub fn pack_to_vec(src: &[u8], level: i32) -> Vec<u8> {
+pub fn pack_to_vec(src: &[u8], level: u8) -> Vec<u8> {
     let mut dst = vec![0u8; brieflz::max_packed_size(src.len())];
     let words = brieflz::workmem_size_level(src.len(), level).unwrap() / 4;
     let mut work = vec![0u32; words.max(1)];
